feat: major codebase refactor again to streamline code and tidyup the vibecoding
