+++ internal/tui/messages.go
@@ -1,0 +1,58 @@
+package tui
+
+import "github.com/christophergyman/claude-quick/internal/devcontainer"
+
+// Message types for async operations in Bubbletea.
+// These are returned by commands and processed in the Update method.
+
+// instancesDiscoveredMsg is sent when project discovery completes
+type instancesDiscoveredMsg struct {
+	instances []devcontainer.ContainerInstance
+}
+
+// instanceStatusRefreshedMsg is sent when container status refresh completes
+type instanceStatusRefreshedMsg struct {
+	statuses []devcontainer.ContainerInstanceWithStatus
+}
+
+// containerStartedMsg is sent when a container finishes starting
+type containerStartedMsg struct{}
+
+// containerErrorMsg is sent when any container operation fails
+type containerErrorMsg struct{ err error }
+
+// tmuxSessionsLoadedMsg is sent when tmux session list is loaded
+type tmuxSessionsLoadedMsg struct{ sessions []string }
+
+// tmuxSessionCreatedMsg is sent when a new tmux session is created
+type tmuxSessionCreatedMsg struct{}
+
+// containerStoppedMsg is sent when a container is stopped
+type containerStoppedMsg struct{}
+
+// containerRestartedMsg is sent when a container is restarted
+type containerRestartedMsg struct{}
+
+// tmuxSessionStoppedMsg is sent when a tmux session is killed
+type tmuxSessionStoppedMsg struct{}
+
+// tmuxSessionRestartedMsg is sent when a tmux session is restarted
+type tmuxSessionRestartedMsg struct{}
+
+// tmuxDetachedMsg is sent when user detaches from tmux
+type tmuxDetachedMsg struct{}
+
+// worktreeCreatedMsg is sent when a new git worktree is created
+type worktreeCreatedMsg struct {
+	worktreePath string
+}
+
+// worktreeDeletedMsg is sent when a git worktree is deleted
+type worktreeDeletedMsg struct{}
+
+// tmuxNotFoundError indicates tmux is not available in the container
+type tmuxNotFoundError struct{}
+
+func (e *tmuxNotFoundError) Error() string {
+	return "tmux not found in container. Please install tmux in your devcontainer."
+}
